Make sure a player moves king out of check
Make sure when a player is in check that
they either block the offending piece
or move their king out of check.

Also, make sure a player can't move
a piece that would result in their king
being in check.

** BAD PRACTICE **
I shouldn't be implementing multiple features
in a single push
******************