Fixed Vulnerable Space Bug
Fixed a bug where if an opposing
player's piece was blocking their
own piece from putting your king in
jeopardy it would still register
as an invalid space.

For instance if a single row
of the board looked like this....
|  | W.K |  |  | B.B |  | B.Q |  |

Before this fix the bishop wouldn't register
as blocking the queen and therefor the king
would be in check.

Now, the Bishop registers as blocking the king!